Improve class organisation
Include `build_command_string` only where necessary.  Make the `query_reply` function part of the base class definition.
Move the `DaliFrame` into a seperate module. This solves a circular import by the helper module. Also, I expect the
`DaliFrame` module to include more functions in the not so distant future.
