+++ dali_interface/hid.py
@@ -16,7 +16,8 @@ from typing import Final, Generator
 
 import usb  # type: ignore
 
-from .dali_interface import DaliFrame, DaliInterface, DaliStatus
+from .dali_interface import DaliInterface
+from .frame import DaliFrame, DaliStatus
 
 logger = logging.getLogger(__name__)
 
@@ -87,9 +88,6 @@ class DaliUsb(DaliInterface):  # pylint: disable=too-many-instance-attributes
         self.receive_sequence_number = 0
         self.last_transmit: int | None = None
 
-        # set the reply timeout to ample amount - timeout event is generated by interface
-        self.reply_timeout = 0.1
-
         logger.debug("try to discover DALI interfaces")
         devices = usb.core.find(find_all=True, idVendor=vendor, idProduct=product)  # pyright: ignore
         # using find_all=True should always return a generator
@@ -141,15 +139,18 @@ class DaliUsb(DaliInterface):  # pylint: disable=too-many-instance-attributes
                 logger.debug("device has integrated power supply")
 
             # read pending messages and discard
-            super().__init__()
             try:
                 while True:
                     self.ep_read.read(getattr(self.ep_read, "wMaxPacketSize"), timeout=10)
                     logger.info("DALI interface - disregard pending messages")
             except Exception:
                 pass
+            super().__init__()
+            # set the reply timeout to ample amount - timeout event is generated by interface
+            self.reply_timeout = 0.1
             return
         # cleanup
+
         self.device = None
         self.ep_read = None
         self.ep_write = None
